option to view readme & view/edit args on first launch
issue #10

Committing here for safekeeping. Seems to work on some initial tests. Since
lots of the script has been touched (especially the handling of maplist) I
need to re-test a bunch of paths, and then add documentation.

Not sure if the preview_readme_and_config option should be true or false by
default in the provided config.