be more aggressive in detecting Quoth requirement
It's unlikely someone would use this word in the readme unless the map requires it, I think? Especially since Quoth is closed-source and can't have things lifted from it for other mods.

Specifically I ran into the mhsp01 readme just saying "Requires Quoth" but I'm sure there are other cases like that too.

In any case, if this triggers a false positive, the new config-preview-edit stuff could be used to correct it.