move bmodel-skipping into maplist generation
The only difference in this approach is that if there is one single legit mapfile that happens to begin with "b_", now we won't launch into that map immediately, while before we would. IMO it's worth risking that situation in order to have a cleaner maplist stored in the gamedir config.