fancier config-editing
Don't automatically enshrine startmap and maplist; just show their current values as examples for what the user might want/need to correct.

Remove examples and a bunch of commentary from the config after it is initially viewed/edited by the user.

Properly handle the user setting a different startmap in the config.