WIP change to manifest v3
Immediate issue is with URL.createObjectURL no longer working/available. Testing a workaround for that, using an "extension resource URL" that doesn't point to any actual file, so I can intercept the fetch event and send the blob in response. Keeping notes about that in issue #26.

Another issue is the lack of DOM parsing.

I'm also not quite sure about the alarms stuff, but haven't gotten to test that.