test fix for "bad" filename characters
Issue #34 : certain filename characters can cause the fetch in the new tab to return not-found.

This means I can't rely on the URL to specify the filename to download, so I'll need to set the filename in the content-disposition header.

Because of this I could really just use any unique string for the URL. I guess I'll use the URI-encoded actual filename... there's a remote chance that it could be useful in future debugging to determine the original filename from that.